Big image loading redesign.
Much "smarter" "on demand" preloading, instead of decoding all images at
startup, do it in the background as we view images and only start
GOMAXPROC goroutines to do this, also set a limit to how many images we
preload to avoid runaway memory usage, will try to be smarter about how
to tune this later.

Also now should be easier to add garbage collection of preloading images
if we start to run out of memory.