Refactoring and remove bad images from list.
Turned canvas into a type with methods instead of just a function with
closures. Add a method to delete images from the list of images, and use
it to delete any images with decode errors, and if the file is gone
after running an !external command.

For this to work had to make the running of exernal commands
synchronous.

Future changes:

- A way to execute external commands in background (although you already
  can do it if you run sh -c ... & or some other such hack.
- Maybe we shouldn't skip images with errors, and instead display the
  error on screen or something.